Add typed-test for various compression types; requires types to be fully known
